refactor: move the download ingestion code to analytics.ingestion
Pure relocation, no behaviour change, ahead of the ingestion rework. The files
move from org.eclipse.openvsx.storage.log to the package layout the rework
targets:

  storage/log/AwsDownloadCountHandler    -> analytics/ingestion/aws/
  storage/log/{CloudFront,Fastly}LogFileParser, LogFileParser, LogRecord
                                         -> analytics/ingestion/aws/
  storage/log/AzureDownloadCountHandler  -> analytics/ingestion/azure/
  storage/log/{DownloadCountProcessor,DownloadCountService}
                                         -> analytics/ingestion/

Class names and file contents are untouched, so every file is a git rename and
the only edits are `package` declarations, the imports that were previously
same-package, and the mock declarations in the tests that name the service.

The one change a compiler could not have caught is in DownloadCountServiceAspect:
its @Around pointcut matches the package as a string, so the move would have
silently stopped the mirror from intercepting isEnabled. That is the reason for
doing this as its own step.
