+++ server/src/main/java/org/eclipse/openvsx/analytics/ingestion/DownloadCountProcessor.java
@@ -7,7 +7,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  ********************************************************************************/
-package org.eclipse.openvsx.storage.log;
+package org.eclipse.openvsx.analytics.ingestion;
 
 import java.time.LocalDateTime;
 import java.util.List;
