+++ server/src/main/java/org/eclipse/openvsx/analytics/ingestion/aws/LogFileParser.java
@@ -10,7 +10,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  *****************************************************************************/
-package org.eclipse.openvsx.storage.log;
+package org.eclipse.openvsx.analytics.ingestion.aws;
 
 import org.jspecify.annotations.Nullable;
