+++ server/src/test/resources/org/eclipse/openvsx/analytics/ingestion/aws/cloudfront.log
