+++ server/src/test/resources/org/eclipse/openvsx/analytics/ingestion/aws/fastly.log
